// src/models/pokemon.models.ts
export class Pokemon {
  Name: string;
  Type: string;
  Height: number;
  Weight: number;
  HP: number;
  Attack: number;
  Defense: number;
  SpecialAttack: number;
  SpecialDefense: number;
  Speed: number;
  Img: string;
  StatsSum: number;
}

export class Generation {
  Id: string;
  NumberOfNewAbilities: number;
}